model_validation_depre/requirements.txt
numpy>=1.23
pandas>=1.5
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
optuna>=3.4
tqdm>=4.65
pyarrow>=12.0
h5py>=3.8
pyyaml>=6.0
joblib>=1.2
torch>=2.1
